requirements.txt
wordcloud~=1.9.4
numpy~=2.1.3
pillow~=11.0.0
matplotlib~=3.9.2